moves the APIs in a folder specific to the recipe
