+++ src/main/java/io/supertokens/webserver/Webserver.java
@@ -23,7 +23,14 @@ import io.supertokens.cliOptions.CLIOptions;
 import io.supertokens.config.Config;
 import io.supertokens.exceptions.QuitProgramException;
 import io.supertokens.output.Logging;
-import io.supertokens.webserver.api.*;
+import io.supertokens.webserver.api.core.ConfigAPI;
+import io.supertokens.webserver.api.core.HelloAPI;
+import io.supertokens.webserver.api.core.NotFoundAPI;
+import io.supertokens.webserver.api.emailpassword.GeneratePasswordResetTokenAPI;
+import io.supertokens.webserver.api.emailpassword.ResetPasswordAPI;
+import io.supertokens.webserver.api.emailpassword.SignInAPI;
+import io.supertokens.webserver.api.emailpassword.SignUpAPI;
+import io.supertokens.webserver.api.session.*;
 import org.apache.catalina.LifecycleException;
 import org.apache.catalina.LifecycleState;
 import org.apache.catalina.connector.Connector;
