+++ src/main/java/io/supertokens/webserver/api/emailpassword/ResetPasswordAPI.java
@@ -14,17 +14,14 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api;
+package io.supertokens.webserver.api.emailpassword;
 
-import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.supertokens.Main;
 import io.supertokens.emailpassword.EmailPassword;
-import io.supertokens.emailpassword.User;
-import io.supertokens.emailpassword.exceptions.WrongCredentialsException;
+import io.supertokens.emailpassword.exceptions.ResetPasswordInvalidTokenException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
-import io.supertokens.utils.Utils;
+import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 
@@ -32,47 +29,57 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
 
-public class SignInAPI extends WebserverAPI {
+public class ResetPasswordAPI extends WebserverAPI {
+    private static final long serialVersionUID = -7529428297450682549L;
 
-    private static final long serialVersionUID = -4641988458637882374L;
-
-    public SignInAPI(Main main) {
+    public ResetPasswordAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/recipe/signin";
+        return "/recipe/user/password/reset";
     }
+
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String email = InputParser.parseStringOrThrowError(input, "email", false);
-        String password = InputParser.parseStringOrThrowError(input, "password", false);
-        assert password != null;
-        assert email != null;
+        String method = InputParser.parseStringOrThrowError(input, "method", false);
+        String token = InputParser.parseStringOrThrowError(input, "token", false);
+        String newPassword = InputParser.parseStringOrThrowError(input, "newPassword", false);
+        assert newPassword != null;
+        assert method != null;
+        assert token != null;
 
-        // logic according to https://github.com/supertokens/supertokens-core/issues/104
+        // logic according to https://github.com/supertokens/supertokens-core/issues/109
 
-        String normalisedEmail = Utils.normaliseEmail(email);
+        if (!method.equals("token")) {
+            throw new ServletException(new WebserverAPI.BadRequestException("Unsupported method for password reset"));
+        }
+
+        if (newPassword.equals("")) {
+            throw new ServletException(
+                    new WebserverAPI.BadRequestException("Password cannot be an empty string"));
+        }
 
         try {
-            User user = EmailPassword.signIn(super.main, normalisedEmail, password);
+            EmailPassword.resetPassword(super.main, token, newPassword);
 
             JsonObject result = new JsonObject();
             result.addProperty("status", "OK");
-            result.add("user", new JsonParser().parse(new Gson().toJson(user)).getAsJsonObject());
             super.sendJsonResponse(200, result, resp);
 
-        } catch (WrongCredentialsException e) {
+        } catch (ResetPasswordInvalidTokenException e) {
             JsonObject result = new JsonObject();
-            result.addProperty("status", "WRONG_CREDENTIALS_ERROR");
+            result.addProperty("status", "RESET_PASSWORD_INVALID_TOKEN_ERROR");
             super.sendJsonResponse(200, result, resp);
-        } catch (StorageQueryException e) {
+        } catch (StorageQueryException | NoSuchAlgorithmException | StorageTransactionLogicException e) {
             throw new ServletException(e);
         }
 
     }
+
 }
