+++ src/main/java/io/supertokens/webserver/api/session/HandshakeAPI.java
@@ -14,71 +14,50 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api;
+package io.supertokens.webserver.api.session;
 
-import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.supertokens.Main;
+import io.supertokens.config.Config;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
-import io.supertokens.session.Session;
 import io.supertokens.session.accessToken.AccessTokenSigningKey;
-import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 
-import javax.crypto.BadPaddingException;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.NoSuchPaddingException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SignatureException;
-import java.security.spec.InvalidKeySpecException;
 
-public class SessionAPI extends WebserverAPI {
-    private static final long serialVersionUID = 7142317017402226537L;
+public class HandshakeAPI extends WebserverAPI {
+    private static final long serialVersionUID = -3647598432179106404L;
 
-    public SessionAPI(Main main) {
+    public HandshakeAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/recipe/session";
+        return "/recipe/handshake";
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String userId = InputParser.parseStringOrThrowError(input, "userId", false);
-        assert userId != null;
-        JsonObject userDataInJWT = InputParser.parseJsonObjectOrThrowError(input, "userDataInJWT", false);
-        assert userDataInJWT != null;
-        JsonObject userDataInDatabase = InputParser.parseJsonObjectOrThrowError(input, "userDataInDatabase", false);
-        assert userDataInDatabase != null;
-
         try {
-            SessionInformationHolder sessionInfo = Session
-                    .createNewSession(main, userId, userDataInJWT, userDataInDatabase);
-
-            JsonObject result = new JsonParser().parse(new Gson().toJson(sessionInfo)).getAsJsonObject();
-
+            JsonObject result = new JsonObject();
             result.addProperty("status", "OK");
             result.addProperty("jwtSigningPublicKey", AccessTokenSigningKey.getInstance(main).getKey().publicKey);
             result.addProperty("jwtSigningPublicKeyExpiryTime",
                     AccessTokenSigningKey.getInstance(main).getKeyExpiryTime());
+            result.addProperty("enableAntiCsrf", Config.getConfig(main).getEnableAntiCSRF());
+            result.addProperty("accessTokenBlacklistingEnabled", Config.getConfig(main).getAccessTokenBlacklisting());
+            result.addProperty("accessTokenValidity", Config.getConfig(main).getAccessTokenValidity());
+            result.addProperty("refreshTokenValidity", Config.getConfig(main).getRefreshTokenValidity());
             super.sendJsonResponse(200, result, resp);
-        } catch (NoSuchAlgorithmException | StorageQueryException | InvalidKeyException | InvalidKeySpecException |
-                StorageTransactionLogicException | SignatureException | IllegalBlockSizeException |
-                BadPaddingException | InvalidAlgorithmParameterException | NoSuchPaddingException e) {
+        } catch (StorageQueryException | StorageTransactionLogicException e) {
             throw new ServletException(e);
         }
     }
-
 }
