+++ src/main/java/io/supertokens/webserver/api/session/RefreshSessionAPI.java
@@ -14,18 +14,17 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api;
+package io.supertokens.webserver.api.session;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.exceptions.TryRefreshTokenException;
+import io.supertokens.exceptions.TokenTheftDetectedException;
 import io.supertokens.exceptions.UnauthorisedException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.session.Session;
-import io.supertokens.session.accessToken.AccessTokenSigningKey;
 import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
@@ -35,39 +34,31 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class VerifySessionAPI extends WebserverAPI {
+public class RefreshSessionAPI extends WebserverAPI {
+    private static final long serialVersionUID = 7142317017402226537L;
 
-    private static final long serialVersionUID = -9169174805902835488L;
-
-    public VerifySessionAPI(Main main) {
+    public RefreshSessionAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/recipe/session/verify";
+        return "/recipe/session/refresh";
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String accessToken = InputParser.parseStringOrThrowError(input, "accessToken", false);
-        assert accessToken != null;
+        String refreshToken = InputParser.parseStringOrThrowError(input, "refreshToken", false);
         String antiCsrfToken = InputParser.parseStringOrThrowError(input, "antiCsrfToken", true);
-        Boolean doAntiCsrfCheck = InputParser.parseBooleanOrThrowError(input, "doAntiCsrfCheck", false);
-        assert doAntiCsrfCheck != null;
+        assert refreshToken != null;
 
         try {
             SessionInformationHolder sessionInfo = Session
-                    .getSession(main, accessToken, antiCsrfToken, doAntiCsrfCheck);
-
+                    .refreshSession(main, refreshToken, antiCsrfToken);
             JsonObject result = new JsonParser().parse(new Gson().toJson(sessionInfo)).getAsJsonObject();
             result.addProperty("status", "OK");
-            result.addProperty("jwtSigningPublicKey", AccessTokenSigningKey.getInstance(main).getKey().publicKey);
-            result.addProperty("jwtSigningPublicKeyExpiryTime",
-                    AccessTokenSigningKey.getInstance(main).getKeyExpiryTime());
             super.sendJsonResponse(200, result, resp);
-
         } catch (StorageQueryException | StorageTransactionLogicException e) {
             throw new ServletException(e);
         } catch (UnauthorisedException e) {
@@ -75,10 +66,15 @@ public class VerifySessionAPI extends WebserverAPI {
             reply.addProperty("status", "UNAUTHORISED");
             reply.addProperty("message", e.getMessage());
             super.sendJsonResponse(200, reply, resp);
-        } catch (TryRefreshTokenException e) {
+        } catch (TokenTheftDetectedException e) {
             JsonObject reply = new JsonObject();
-            reply.addProperty("status", "TRY_REFRESH_TOKEN");
-            reply.addProperty("message", e.getMessage());
+            reply.addProperty("status", "TOKEN_THEFT_DETECTED");
+
+            JsonObject session = new JsonObject();
+            session.addProperty("handle", e.sessionHandle);
+            session.addProperty("userId", e.userId);
+            reply.add("session", session);
+
             super.sendJsonResponse(200, reply, resp);
         }
     }
