+++ src/main/java/io/supertokens/webserver/api/session/SessionAPI.java
@@ -14,68 +14,71 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api;
+package io.supertokens.webserver.api.session;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.exceptions.TokenTheftDetectedException;
-import io.supertokens.exceptions.UnauthorisedException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.session.Session;
+import io.supertokens.session.accessToken.AccessTokenSigningKey;
 import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 
+import javax.crypto.BadPaddingException;
+import javax.crypto.IllegalBlockSizeException;
+import javax.crypto.NoSuchPaddingException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.security.InvalidAlgorithmParameterException;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
+import java.security.SignatureException;
+import java.security.spec.InvalidKeySpecException;
 
-public class RefreshSessionAPI extends WebserverAPI {
+public class SessionAPI extends WebserverAPI {
     private static final long serialVersionUID = 7142317017402226537L;
 
-    public RefreshSessionAPI(Main main) {
+    public SessionAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/recipe/session/refresh";
+        return "/recipe/session";
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String refreshToken = InputParser.parseStringOrThrowError(input, "refreshToken", false);
-        String antiCsrfToken = InputParser.parseStringOrThrowError(input, "antiCsrfToken", true);
-        assert refreshToken != null;
+        String userId = InputParser.parseStringOrThrowError(input, "userId", false);
+        assert userId != null;
+        JsonObject userDataInJWT = InputParser.parseJsonObjectOrThrowError(input, "userDataInJWT", false);
+        assert userDataInJWT != null;
+        JsonObject userDataInDatabase = InputParser.parseJsonObjectOrThrowError(input, "userDataInDatabase", false);
+        assert userDataInDatabase != null;
 
         try {
             SessionInformationHolder sessionInfo = Session
-                    .refreshSession(main, refreshToken, antiCsrfToken);
+                    .createNewSession(main, userId, userDataInJWT, userDataInDatabase);
+
             JsonObject result = new JsonParser().parse(new Gson().toJson(sessionInfo)).getAsJsonObject();
+
             result.addProperty("status", "OK");
+            result.addProperty("jwtSigningPublicKey", AccessTokenSigningKey.getInstance(main).getKey().publicKey);
+            result.addProperty("jwtSigningPublicKeyExpiryTime",
+                    AccessTokenSigningKey.getInstance(main).getKeyExpiryTime());
             super.sendJsonResponse(200, result, resp);
-        } catch (StorageQueryException | StorageTransactionLogicException e) {
+        } catch (NoSuchAlgorithmException | StorageQueryException | InvalidKeyException | InvalidKeySpecException |
+                StorageTransactionLogicException | SignatureException | IllegalBlockSizeException |
+                BadPaddingException | InvalidAlgorithmParameterException | NoSuchPaddingException e) {
             throw new ServletException(e);
-        } catch (UnauthorisedException e) {
-            JsonObject reply = new JsonObject();
-            reply.addProperty("status", "UNAUTHORISED");
-            reply.addProperty("message", e.getMessage());
-            super.sendJsonResponse(200, reply, resp);
-        } catch (TokenTheftDetectedException e) {
-            JsonObject reply = new JsonObject();
-            reply.addProperty("status", "TOKEN_THEFT_DETECTED");
-
-            JsonObject session = new JsonObject();
-            session.addProperty("handle", e.sessionHandle);
-            session.addProperty("userId", e.userId);
-            reply.add("session", session);
-
-            super.sendJsonResponse(200, reply, resp);
         }
     }
+
 }
