+++ src/main/java/io/supertokens/webserver/api/session/VerifySessionAPI.java
@@ -14,13 +14,19 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api;
+package io.supertokens.webserver.api.session;
 
+import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import io.supertokens.Main;
+import io.supertokens.exceptions.TryRefreshTokenException;
 import io.supertokens.exceptions.UnauthorisedException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
+import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.session.Session;
+import io.supertokens.session.accessToken.AccessTokenSigningKey;
+import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.webserver.InputParser;
 import io.supertokens.webserver.WebserverAPI;
 
@@ -29,64 +35,51 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class SessionDataAPI extends WebserverAPI {
-    private static final long serialVersionUID = -6901312482713647177L;
+public class VerifySessionAPI extends WebserverAPI {
 
-    public SessionDataAPI(Main main) {
+    private static final long serialVersionUID = -9169174805902835488L;
+
+    public VerifySessionAPI(Main main) {
         super(main);
     }
 
     @Override
     public String getPath() {
-        return "/recipe/session/data";
+        return "/recipe/session/verify";
     }
 
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        String sessionHandle = InputParser.getQueryParamOrThrowError(req, "sessionHandle", false);
-        assert sessionHandle != null;
-
-        try {
-            JsonObject userDataInDatabase = Session.getSessionData(main, sessionHandle);
-
-            JsonObject result = new JsonObject();
-            result.addProperty("status", "OK");
-            result.add("userDataInDatabase", userDataInDatabase);
-            super.sendJsonResponse(200, result, resp);
-
-        } catch (StorageQueryException e) {
-            throw new ServletException(e);
-        } catch (UnauthorisedException e) {
-            JsonObject reply = new JsonObject();
-            reply.addProperty("status", "UNAUTHORISED");
-            reply.addProperty("message", e.getMessage());
-            super.sendJsonResponse(200, reply, resp);
-        }
-    }
-
-    @Override
-    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
-        String sessionHandle = InputParser.parseStringOrThrowError(input, "sessionHandle", false);
-        assert sessionHandle != null;
-        JsonObject userDataInDatabase = InputParser.parseJsonObjectOrThrowError(input, "userDataInDatabase", false);
-        assert userDataInDatabase != null;
+        String accessToken = InputParser.parseStringOrThrowError(input, "accessToken", false);
+        assert accessToken != null;
+        String antiCsrfToken = InputParser.parseStringOrThrowError(input, "antiCsrfToken", true);
+        Boolean doAntiCsrfCheck = InputParser.parseBooleanOrThrowError(input, "doAntiCsrfCheck", false);
+        assert doAntiCsrfCheck != null;
 
         try {
-            Session.updateSession(main, sessionHandle, userDataInDatabase, null, null);
+            SessionInformationHolder sessionInfo = Session
+                    .getSession(main, accessToken, antiCsrfToken, doAntiCsrfCheck);
 
-            JsonObject result = new JsonObject();
+            JsonObject result = new JsonParser().parse(new Gson().toJson(sessionInfo)).getAsJsonObject();
             result.addProperty("status", "OK");
+            result.addProperty("jwtSigningPublicKey", AccessTokenSigningKey.getInstance(main).getKey().publicKey);
+            result.addProperty("jwtSigningPublicKeyExpiryTime",
+                    AccessTokenSigningKey.getInstance(main).getKeyExpiryTime());
             super.sendJsonResponse(200, result, resp);
 
-        } catch (StorageQueryException e) {
+        } catch (StorageQueryException | StorageTransactionLogicException e) {
             throw new ServletException(e);
         } catch (UnauthorisedException e) {
             JsonObject reply = new JsonObject();
             reply.addProperty("status", "UNAUTHORISED");
             reply.addProperty("message", e.getMessage());
             super.sendJsonResponse(200, reply, resp);
+        } catch (TryRefreshTokenException e) {
+            JsonObject reply = new JsonObject();
+            reply.addProperty("status", "TRY_REFRESH_TOKEN");
+            reply.addProperty("message", e.getMessage());
+            super.sendJsonResponse(200, reply, resp);
         }
     }
-
 }
